feat: add conversation filter bar
