+++ src/components-next/common/filters/index.ts
@@ -1,2 +1,3 @@
 export * from './FilterBar';
 export * from './FilterButton';
+export * from './ConversationFilterBar';
